requirement.txt
google-genai
openai
python-dotenv
streamlit